Changed AircraftManagementDatabase to properly add a flight descriptor to a management record and to make the getStatus method return a String for the labels in GOC.
Added a constructor that will create a new gate array to make new gates due to previously getting null pointer exceptions.

Added a new method in ManagementRecord to get the status using a switch with the mCode to find the correct status of the record.

Added a constructor to make a new ArrayList for details.