Changed AircraftManagementDatabase to properly add a flight descriptor to a management record and to make the getStatus method return a String for the labels in GOC.
Added a constructor that will create a new gate array to make new gates due to previously getting null pointer exceptions.

Added a new method in ManagementRecord to get the status using a switch with the mCode to find the correct status of the record.

Added a constructor to make a new ArrayList for details.

Implemented GOC, LATC & PublicInfo.

Added observer methods to setStatus in AircraftManagementDatabase.

Created data for the AircraftManagementDatabase to test the Controllers and Views.

Changed the radarDetect so that it checks to see if the to of itinerary is Stirling so that it sets the status to WAITING_TO_LAND. Added a toString as well for the views to see the flightcode of the flight and where is from, where is the airplane landing and where its going next if applicable.

Added a getter for details in PassengerList for getting the size of the ArrayList for amount of passengers.